requirements.txt
gradio
pandas
numpy
scikit-learn
matplotlib
ydata-profiling
openai
python-dotenv